refactor: move stream code into its own module
Issue STOR-212
